# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "Bakery APIs Document",
        "description": "These API-s help a bakery manage their products and serve orders",
        "termsOfService": "",
        "contact": {
            "name": "Paige Turner",
            "email": "paige.turner@gmail.com",
            "url": "https://paigeturner.any"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Local server"
        }
    ],
...
    "paths": {
...
        "/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "description": "Places a new order.",
                "operationId": "placeOrder",
                "requestBody": {
                    "description": "Order Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful place order response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/OrderResult"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Result"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/definitions/Result"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Result": {
            "type": "object",
            "required": ["message", "code"],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "Operation result"
                },
                "code": {
                    "type": "string",
                    "description": "Product Code"
                }
            }
        },
        "ProductOrder": {
            "type": "object",
            "required": ["count", "code"],
            "properties": {
                "count": {
                    "type": "number",
                    "description": "Packaging Option Count"
                },
                "code": {
                    "type": "number",
                    "description": "Product Code"
                }
            }
        },
        "Order": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/ProductOrder"
            }
        },
        "Packaging": {
            "type": "object",
            "required": ["packs", "count", "price"],
            "properties": {
                "count": {
                    "type": "number",
                    "description": "Number of items in the package"
                },
                "packs": {
                    "type": "number",
                    "description": "Number of packs"
                },
                "price": {
                    "type": "number",
                    "description": "Price of one package"
                }
            }
        },
        "ProductPackaging": {
            "code": {
                "type": "number",
                "description": "Product Code"
            },
            "count": {
                "type": "number",
                "description": "Packaging Option Count"
            },
            "totalPrice": {
                "type": "number",
                "description": "Total price for this product order"
            },
            "packaging": {
                "type": "array",
                "items": {
                    "$ref": "#/definitions/Packaging"
                }
            }
        },
        "OrderResult": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/ProductPackaging"
            }
        }
    }
}